backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<3.16
djangorestframework-simplejwt[crypto]>=5.3
django-cors-headers>=4.4
